fix:delete document upload service and refactor the logic
