+++ src/app/view/admin/add-users/add-users.component.ts
@@ -69,7 +69,7 @@ export class AddUsersComponent {
       ]),
       password: new FormControl(null),
       confirmPassword: new FormControl(null),
-      termsAndConditions: new FormControl(false)
+      termsAndConditions: new FormControl(false),
     });
   }
   emaiErrors() {
